pset8/jar/test_jar.py
from jar import Jar  # Assuming the Jar class is defined in a module named 'jar'

# Test case for the initialization of the Jar class
def test_init():
    jar = Jar()
    # Check if the default capacity is set to 12
    assert jar.capacity == 12
    # Check if the initial size is 0
    assert jar.size == 0

# Test case for the string representation of the Jar class
def test_str():
    jar = Jar()
    # Check if an empty jar is represented as an empty string
    assert str(jar) == ""
    # Deposit 1 cookie and check the string representation
    jar.deposit(1)
    assert str(jar) == "🍪"
    # Deposit 11 more cookies and check the string representation
    jar.deposit(11)
    assert str(jar) == "🍪🍪🍪🍪🍪🍪🍪🍪🍪🍪🍪🍪"

# Test case for depositing cookies into the Jar class
def test_deposit():
    jar = Jar()
    # Deposit 5 cookies and check if the size is updated
    jar.deposit(5)
    assert jar.size == 5
    # Deposit 7 more cookies, exceeding the capacity, and check if the size is capped at the capacity
    jar.deposit(7)
    assert jar.size == 12
    # Check if the string representation reflects the deposited cookies
    assert str(jar) == "🍪🍪🍪🍪🍪🍪🍪🍪🍪🍪🍪🍪"
    # Test depositing more than the capacity and check if it raises a ValueError
    try:
        jar.deposit(1)
    except ValueError as e:
        assert str(e) == "Exceeds the cookie jar's capacity"

# Test case for withdrawing cookies from the Jar class
def test_withdraw():
    jar = Jar()
    # Deposit 10 cookies and withdraw 5, checking if the size and string representation are updated
    jar.deposit(10)
    jar.withdraw(5)
    assert jar.size == 5
    assert str(jar) == "🍪🍪🍪🍪🍪"
    # Test withdrawing more cookies than available and check if it raises a ValueError
    try:
        jar.withdraw(7)
    except ValueError as e:
        assert str(e) == "Not enough cookies in the jar"

"""
This code defines a set of test cases for a hypothetical Jar class, presumably defined in a module named 'jar'. 
The first test case (test_init()) checks if the initialization of a Jar object sets its default capacity to 12 and the initial size to 0. 
The second test case (test_str()) verifies the string representation of the jar, 
starting with an empty jar and then depositing and checking different quantities of cookies. 
The third test case (test_deposit()) focuses on the deposit functionality, 
checking if the size is updated correctly and if exceeding the capacity raises a ValueError with the appropriate message. 
The fourth test case (test_withdraw()) tests the withdrawal functionality, 
ensuring that the size and string representation are updated accordingly and that attempting to withdraw more cookies than available 
raises a ValueError with the expected message. The tests collectively assess the basic functionalities and error handling of the Jar class.
"""